requirements.txt
pdfminer.six
spacy
scikit-learn
nltk
pandas
numpy
